extract structures from near-network into -primitives
