+++ chain/client-primitives/Cargo.toml
@@ -18,8 +18,8 @@ near-chain-configs = { path = "../../core/chain-configs" }
 
 near-chunks-primitives = { path = "../chunks-primitives" }
 near-crypto = { path = "../../core/crypto" }
-near-network = { path = "../network" }
+near-network-primitives = { path = "../network-primitives" }
 near-primitives = { path = "../../core/primitives" }
 
 [features]
-metric_recorder = []
+metric_recorder = ["near-network-primitives/metric_recorder"]
