+++ chain/network-primitives/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "near-network-primitives"
+version = "0.1.0"
+authors = ["Near Inc <hello@nearprotocol.com>"]
+edition = "2018"
+
+[dependencies]
+actix = "=0.11.0-beta.2"
+tokio = { version = "1.1", features = ["full"] }
+futures = "0.3"
+chrono = { version = "0.4.4", features = ["serde"] }
+borsh = "0.8.1"
+serde = { version = "1", features = [ "derive" ] }
+serde_json = "1"
+strum = { version = "0.20", features = ["derive"] }
+tracing = "0.1.13"
+conqueue = "0.4.0"
+bytes = "1"
+bytesize = "1.0.1"
+tokio-util = { version = "0.6", features = ["codec"] }
+byteorder = "1.2"
+lazy_static = "1.4"
+near-rust-allocator-proxy = "0.2.9"
+
+near-crypto = { path = "../../core/crypto" }
+near-primitives = { path = "../../core/primitives" }
+near-performance-metrics = { path = "../../utils/near-performance-metrics" }
+near-metrics = { path = "../../core/metrics" }
+
+[features]
+adversarial = []
+metric_recorder = []
+sandbox = []
+performance_stats = ["near-performance-metrics/performance_stats"]
+# --
+__internal_primitives = []
