+++ chain/network-primitives/src/lib.rs
@@ -1,0 +1,15 @@
+#[cfg(feature = "__internal_primitives")]
+pub mod cache;
+#[cfg(not(feature = "__internal_primitives"))]
+mod cache;
+#[cfg(feature = "__internal_primitives")]
+pub mod codec;
+#[cfg(not(feature = "__internal_primitives"))]
+mod codec;
+
+#[cfg(feature = "metric_recorder")]
+pub mod recorder;
+
+pub mod metrics;
+pub mod routing;
+pub mod types;
