+++ chain/network/Cargo.toml
@@ -5,7 +5,6 @@ authors = ["Near Inc <hello@nearprotocol.com>"]
 edition = "2018"
 
 [dependencies]
-bytes = "1"
 actix = "=0.11.0-beta.2"
 tokio = { version = "1.1", features = ["full"] }
 tokio-util = { version = "0.6", features = ["codec"] }
@@ -13,31 +12,29 @@ tokio-stream = { version = "0.1.2", features = ["net"] }
 futures = "0.3"
 chrono = { version = "0.4.4", features = ["serde"] }
 serde = { version = "1", features = [ "derive" ] }
-serde_json = "1"
 rand = "0.7"
-byteorder = "1.2"
 lazy_static = "1.4"
 tracing = "0.1.13"
 strum = { version = "0.20", features = ["derive"] }
 near-rust-allocator-proxy = "0.2.9"
-bytesize = "1.0.1"
 conqueue = "0.4.0"
 
 borsh = "0.8.1"
 cached = "0.23"
 
+near-network-primitives = { path = "../network-primitives", features = ["__internal_primitives"] }
 near-chain-configs = { path = "../../core/chain-configs" }
 near-crypto = { path = "../../core/crypto" }
 near-primitives = { path = "../../core/primitives" }
 near-store = { path = "../../core/store" }
 near-metrics = { path = "../../core/metrics" }
-near-chain = { path = "../chain" }
 near-performance-metrics = { path = "../../utils/near-performance-metrics" }
 near-performance-metrics-macros = { path = "../../utils/near-performance-metrics-macros" }
 
 delay-detector = { path = "../../tools/delay_detector", optional = true}
 
 [dev-dependencies]
+near-chain = { path = "../chain" }
 near-logger-utils = {path = "../../test-utils/logger"}
 near-actix-test-utils = { path = "../../test-utils/actix-test-utils" }
 tempfile = "3"
@@ -47,11 +44,11 @@ near-client = { path = "../client" }
 near-telemetry = { path = "../telemetry" }
 
 [features]
-adversarial = []
+adversarial = ["near-network-primitives/adversarial"]
 metric_recorder = []
 delay_detector = ["delay-detector"]
-performance_stats = ["near-performance-metrics/performance_stats"]
-sandbox = []
+performance_stats = ["near-network-primitives/performance_stats", "near-performance-metrics/performance_stats"]
+sandbox = ["near-network-primitives/sandbox"]
 
 [[bench]]
 name = "graph"
