+++ chain/network/src/peer.rs
@@ -26,7 +26,6 @@ use near_primitives::version::{
     ProtocolVersion, OLDEST_BACKWARD_COMPATIBLE_PROTOCOL_VERSION, PROTOCOL_VERSION,
 };
 
-use crate::codec::{self, bytes_to_peer_message, peer_message_to_bytes, Codec};
 use crate::rate_counter::RateCounter;
 #[cfg(feature = "metric_recorder")]
 use crate::recorder::{PeerMessageMetadata, Status};
@@ -41,10 +40,11 @@ use crate::types::{
     UPDATE_INTERVAL_LAST_TIME_RECEIVED_MESSAGE,
 };
 use crate::PeerManagerActor;
-use crate::{metrics, NetworkResponses};
+use crate::{metrics::NetworkMetrics, NetworkResponses};
 #[cfg(feature = "delay_detector")]
 use delay_detector::DelayDetector;
-use metrics::NetworkMetrics;
+use near_network_primitives::codec::{self, bytes_to_peer_message, peer_message_to_bytes, Codec};
+use near_network_primitives::metrics;
 use near_performance_metrics_macros::perf;
 use near_primitives::sharding::PartialEncodedChunk;
 use near_rust_allocator_proxy::allocator::get_tid;
